fix: resolve release build crash + type mismatches in FRB bindings
- Fix AuthCodeResult mirror types: code_length u32, resend_after_seconds u64
- Make check_connection async to avoid Tokio reactor panic in release builds
- Remove unused imports from onboarding, settings, upload screens
- Update .gitignore: exclude jniLibs and screenshots from tracking
- Regenerate FRB bindings with corrected types
- Verified: APK launches clean on emulator (x86_64), no runtime errors

+++ app_flutter/lib/src/rust/api/mirror.dart
@@ -121,7 +121,7 @@ class AppSettings {
 class AuthCodeResult {
   final String status;
   final int? codeLength;
-  final int? resendAfterSeconds;
+  final BigInt? resendAfterSeconds;
   final String? delivery;
 
   const AuthCodeResult({
